// admin/src/gen/types/UpdateUserRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UpdateUserRequest = {
  /**
   * @type array | undefined
   */
  badge_ids?: string[]
  /**
   * @type string | undefined
   */
  description?: string
  /**
   * @type string | undefined
   */
  first_name?: string
  /**
   * @type string | undefined
   */
  last_name?: string
  /**
   * @type string | undefined
   */
  location_id?: string
  /**
   * @type array | undefined
   */
  opportunity_ids?: string[]
  /**
   * @type string | undefined
   */
  title?: string
}